// src/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
};

/** A single film. */
export type Film = Node & {
  __typename?: "Film";
  characterConnection?: Maybe<FilmCharactersConnection>;
  /** The ISO 8601 date format of the time that this resource was created. */
  created?: Maybe<Scalars["String"]["output"]>;
  /** The name of the director of this film. */
  director?: Maybe<Scalars["String"]["output"]>;
  /** The ISO 8601 date format of the time that this resource was edited. */
  edited?: Maybe<Scalars["String"]["output"]>;
  /** The episode number of this film. */
  episodeID?: Maybe<Scalars["Int"]["output"]>;
  /** The ID of an object */
  id: Scalars["ID"]["output"];
  /** The opening paragraphs at the beginning of this film. */
  openingCrawl?: Maybe<Scalars["String"]["output"]>;
  planetConnection?: Maybe<FilmPlanetsConnection>;
  /** The name(s) of the producer(s) of this film. */
  producers?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>;
  /** The ISO 8601 date format of film release at original creator country. */
  releaseDate?: Maybe<Scalars["String"]["output"]>;
  speciesConnection?: Maybe<FilmSpeciesConnection>;
  starshipConnection?: Maybe<FilmStarshipsConnection>;
  /** The title of this film. */
  title?: Maybe<Scalars["String"]["output"]>;
  vehicleConnection?: Maybe<FilmVehiclesConnection>;
};
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars["ID"]["output"];
};

/** Information about pagination in a connection. */
export type PageInfo = {
  __typename?: "PageInfo";
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars["String"]["output"]>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars["Boolean"]["output"];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars["Boolean"]["output"];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type Person = Node & {
  __typename?: "Person";
  /**
   * The birth year of the person, using the in-universe standard of BBY or ABY -
   * Before the Battle of Yavin or After the Battle of Yavin. The Battle of Yavin is
   * a battle that occurs at the end of Star Wars episode IV: A New Hope.
   */
  birthYear?: Maybe<Scalars["String"]["output"]>;
  /** The ISO 8601 date format of the time that this resource was created. */
  created?: Maybe<Scalars["String"]["output"]>;
  /** The ISO 8601 date format of the time that this resource was edited. */
  edited?: Maybe<Scalars["String"]["output"]>;
  /**
   * The eye color of this person. Will be "unknown" if not known or "n/a" if the
   * person does not have an eye.
   */
  eyeColor?: Maybe<Scalars["String"]["output"]>;
  filmConnection?: Maybe<PersonFilmsConnection>;
  /**
   * The gender of this person. Either "Male", "Female" or "unknown",
   * "n/a" if the person does not have a gender.
   */
  gender?: Maybe<Scalars["String"]["output"]>;
  /**
   * The hair color of this person. Will be "unknown" if not known or "n/a" if the
   * person does not have hair.
   */
  hairColor?: Maybe<Scalars["String"]["output"]>;
  /** The height of the person in centimeters. */
  height?: Maybe<Scalars["Int"]["output"]>;
  /** A planet that this person was born on or inhabits. */
  homeworld?: Maybe<Planet>;
  /** The ID of an object */
  id: Scalars["ID"]["output"];
  /** The mass of the person in kilograms. */
  mass?: Maybe<Scalars["Float"]["output"]>;
  /** The name of this person. */
  name?: Maybe<Scalars["String"]["output"]>;
  /** The skin color of this person. */
  skinColor?: Maybe<Scalars["String"]["output"]>;
  /** The species that this person belongs to, or null if unknown. */
  species?: Maybe<Species>;
  starshipConnection?: Maybe<PersonStarshipsConnection>;
  vehicleConnection?: Maybe<PersonVehiclesConnection>;
};
// ...
export type PersonFilmsConnection = {
  __typename?: "PersonFilmsConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<PersonFilmsEdge>>>;
  /**
   * A list of all of the objects returned in the connection. This is a convenience
   * field provided for quickly exploring the API; rather than querying for
   * "{ edges { node } }" when no edge data is needed, this field can be be used
   * instead. Note that when clients like Relay need to fetch the "cursor" field on
   * the edge to enable efficient pagination, this shortcut cannot be used, and the
   * full "{ edges { node } }" version should be used instead.
   */
  films?: Maybe<Array<Maybe<Film>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /**
   * A count of the total number of objects in this connection, ignoring pagination.
   * This allows a client to fetch the first five objects by passing "5" as the
   * argument to "first", then fetch the total count so it could display "5 of 83",
   * for example.
   */
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};

/** An edge in a connection. */
export type PersonFilmsEdge = {
  __typename?: "PersonFilmsEdge";
  /** A cursor for use in pagination */
  cursor: Scalars["String"]["output"];
  /** The item at the end of the edge */
  node?: Maybe<Film>;
};

/** A connection to a list of items. */
export type PersonStarshipsConnection = {
  __typename?: "PersonStarshipsConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<PersonStarshipsEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /**
   * A list of all of the objects returned in the connection. This is a convenience
   * field provided for quickly exploring the API; rather than querying for
   * "{ edges { node } }" when no edge data is needed, this field can be be used
   * instead. Note that when clients like Relay need to fetch the "cursor" field on
   * the edge to enable efficient pagination, this shortcut cannot be used, and the
   * full "{ edges { node } }" version should be used instead.
   */
  starships?: Maybe<Array<Maybe<Starship>>>;
  /**
   * A count of the total number of objects in this connection, ignoring pagination.
   * This allows a client to fetch the first five objects by passing "5" as the
   * argument to "first", then fetch the total count so it could display "5 of 83",
   * for example.
   */
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};

/** An edge in a connection. */
export type PersonStarshipsEdge = {
  __typename?: "PersonStarshipsEdge";
  /** A cursor for use in pagination */
  cursor: Scalars["String"]["output"];
  /** The item at the end of the edge */
  node?: Maybe<Starship>;
};

/** A connection to a list of items. */
export type PersonVehiclesConnection = {
  __typename?: "PersonVehiclesConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<PersonVehiclesEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /**
   * A count of the total number of objects in this connection, ignoring pagination.
   * This allows a client to fetch the first five objects by passing "5" as the
   * argument to "first", then fetch the total count so it could display "5 of 83",
   * for example.
   */
  totalCount?: Maybe<Scalars["Int"]["output"]>;
  /**
   * A list of all of the objects returned in the connection. This is a convenience
   * field provided for quickly exploring the API; rather than querying for
   * "{ edges { node } }" when no edge data is needed, this field can be be used
   * instead. Note that when clients like Relay need to fetch the "cursor" field on
   * the edge to enable efficient pagination, this shortcut cannot be used, and the
   * full "{ edges { node } }" version should be used instead.
   */
  vehicles?: Maybe<Array<Maybe<Vehicle>>>;
};

/** An edge in a connection. */
export type PersonVehiclesEdge = {
  __typename?: "PersonVehiclesEdge";
  /** A cursor for use in pagination */
  cursor: Scalars["String"]["output"];
  /** The item at the end of the edge */
  node?: Maybe<Vehicle>;
};

/**
 * A large mass, planet or planetoid in the Star Wars Universe, at the time of
 * 0 ABY.
 */
export type Planet = Node & {
  __typename?: "Planet";
  /** The climates of this planet. */
  climates?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>;
  /** The ISO 8601 date format of the time that this resource was created. */
  created?: Maybe<Scalars["String"]["output"]>;
  /** The diameter of this planet in kilometers. */
  diameter?: Maybe<Scalars["Int"]["output"]>;
  /** The ISO 8601 date format of the time that this resource was edited. */
  edited?: Maybe<Scalars["String"]["output"]>;
  filmConnection?: Maybe<PlanetFilmsConnection>;
  /**
   * A number denoting the gravity of this planet, where "1" is normal or 1 standard
   * G. "2" is twice or 2 standard Gs. "0.5" is half or 0.5 standard Gs.
   */
  gravity?: Maybe<Scalars["String"]["output"]>;
  /** The ID of an object */
  id: Scalars["ID"]["output"];
  /** The name of this planet. */
  name?: Maybe<Scalars["String"]["output"]>;
  /**
   * The number of standard days it takes for this planet to complete a single orbit
   * of its local star.
   */
  orbitalPeriod?: Maybe<Scalars["Int"]["output"]>;
  /** The average population of sentient beings inhabiting this planet. */
  population?: Maybe<Scalars["Float"]["output"]>;
  residentConnection?: Maybe<PlanetResidentsConnection>;
  /**
   * The number of standard hours it takes for this planet to complete a single
   * rotation on its axis.
   */
  rotationPeriod?: Maybe<Scalars["Int"]["output"]>;
  /**
   * The percentage of the planet surface that is naturally occurring water or bodies
   * of water.
   */
  surfaceWater?: Maybe<Scalars["Float"]["output"]>;
  /** The terrains of this planet. */
  terrains?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>;
};
// ...
export type PlanetFilmsConnection = {
  __typename?: "PlanetFilmsConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<PlanetFilmsEdge>>>;
  /**
   * A list of all of the objects returned in the connection. This is a convenience
   * field provided for quickly exploring the API; rather than querying for
   * "{ edges { node } }" when no edge data is needed, this field can be be used
   * instead. Note that when clients like Relay need to fetch the "cursor" field on
   * the edge to enable efficient pagination, this shortcut cannot be used, and the
   * full "{ edges { node } }" version should be used instead.
   */
  films?: Maybe<Array<Maybe<Film>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /**
   * A count of the total number of objects in this connection, ignoring pagination.
   * This allows a client to fetch the first five objects by passing "5" as the
   * argument to "first", then fetch the total count so it could display "5 of 83",
   * for example.
   */
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export type Species = Node & {
  __typename?: "Species";
  /** The average height of this species in centimeters. */
  averageHeight?: Maybe<Scalars["Float"]["output"]>;
  /** The average lifespan of this species in years, null if unknown. */
  averageLifespan?: Maybe<Scalars["Int"]["output"]>;
  /** The classification of this species, such as "mammal" or "reptile". */
  classification?: Maybe<Scalars["String"]["output"]>;
  /** The ISO 8601 date format of the time that this resource was created. */
  created?: Maybe<Scalars["String"]["output"]>;
  /** The designation of this species, such as "sentient". */
  designation?: Maybe<Scalars["String"]["output"]>;
  /** The ISO 8601 date format of the time that this resource was edited. */
  edited?: Maybe<Scalars["String"]["output"]>;
  /**
   * Common eye colors for this species, null if this species does not typically
   * have eyes.
   */
  eyeColors?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>;
  filmConnection?: Maybe<SpeciesFilmsConnection>;
  /**
   * Common hair colors for this species, null if this species does not typically
   * have hair.
   */
  hairColors?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>;
  /** A planet that this species originates from. */
  homeworld?: Maybe<Planet>;
  /** The ID of an object */
  id: Scalars["ID"]["output"];
  /** The language commonly spoken by this species. */
  language?: Maybe<Scalars["String"]["output"]>;
  /** The name of this species. */
  name?: Maybe<Scalars["String"]["output"]>;
  personConnection?: Maybe<SpeciesPeopleConnection>;
  /**
   * Common skin colors for this species, null if this species does not typically
   * have skin.
   */
  skinColors?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>;
};